backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
redis==5.0.4
pandas==2.1.4
openpyxl==3.1.2
numpy==1.26.4
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
# Optionnels (IA avancée)
tensorflow==2.16.1
xgboost==2.0.3
scikit-learn==1.4.2